Move index fixture into conftest.py file
So that it can be used by all tests in tests/h/search.
